refactor(corsa): move relative module resolution to a submodule
Adding the TS2322 ArrayBuffer/SharedArrayBuffer suppression rule pushed diagnostics.rs past its source-length budget, failing the test-scripts length check. Move the cohesive relative-module-resolution helpers and their test into diagnostics/module_resolution.rs, re-exporting relative_module_resolves_on_disk so the diagnostics module's public surface (consumed by cli.rs) is unchanged.
